Leave request implementation

+++ Dtos/LeaveBalance/CreateLeaveBalanceDto.cs
@@ -2,9 +2,9 @@
 {
     public class CreateLeaveBalanceDto
     {
-        public int Annual { get; set; }
+        public double Annual { get; set; }
 
-        public int Sick { get; set; }
+        public double Sick { get; set; }
 
         public Guid UserId { get; set; }
     }
